+++ spec/mongoid/errors/readonly_attribute_spec.rb
@@ -1,0 +1,29 @@
+require "spec_helper"
+
+describe Mongoid::Errors::ReadonlyAttribute do
+
+  describe "#message" do
+
+    let(:error) do
+      described_class.new(:title, "mr")
+    end
+
+    it "contains the problem in the message" do
+      error.message.should include(
+        "Attempted to set the readonly attribute 'title' with the value: mr."
+      )
+    end
+
+    it "contains the summary in the message" do
+      error.message.should include(
+        "Attributes flagged as readonly via Model.attr_readonly can only"
+      )
+    end
+
+    it "contains the resolution in the message" do
+      error.message.should include(
+        "Don't define 'title' as readonly, or do not attempt to update"
+      )
+    end
+  end
+end